Created fallback for missing config values
The script will now use the default config values for any missing config keys and notify the user as it does so.

This still won't help if you put a wrongly formatted value in the config file, though. That's on you.

Also fixed an erroneous key name in the default config.